make spec / implementation / faq separate pages.  goal is to make the site less mouseover-ish.  for a specification, it seems to me one should easily be able to download or cut and paste it, so that is sort of the goal.
note the leaf pages don't have headers or anything.  not as formal but in the spirit
of something like:

  https://www.ietf.org/rfc/rfc4627.txt

if there is a consensus to put them in, go ahead though.

as text was in tooltips i had to pick some new ordering.  some is in the bottom in "comments".
in general i have attempted not to edit or change text as this is a standard -- although i have
removed a redundant word here or there.